Collision and adjustments
Implemented collision, need some tweaks to offsets. But it is on the
way!
Also did some adjustments to the drawmanager (in preparation for the
view and viewports), also created player entity, and fixed some memory
leaks and other minor adjustments to several classes including (but not
limited to) scenes, entities and collision.